% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildMultiBlock.R
\name{BuildMultiBlock}
\alias{BuildMultiBlock}
\title{BuildMultiBlock
Creates a MultiBlock object}
\usage{
BuildMultiBlock(..., ignore.names = FALSE, ignore.size = FALSE)
}
\arguments{
\item{...}{Data.frames, matrices, or MultiBlocks to be combined.}

\item{ignore.names}{The MultiBlock is created without checking the sample names. Names are replaced by integers. Useful when the nomenclature to name samples is not consistent across blocks but there exist sample correspondence across blocks.}

\item{ignore.size}{The MultiBlock is created without checking the sample size. This is set as TRUE when dealing with replicate data and we want to consider each replicate batch as a stand alone block. See SplitRW().}
}
\value{
The MultiBlock.
}
\description{
BuildMultiBlock
Creates a MultiBlock object
}
\examples{
 b1 = matrix(rnorm(500),10,50) # 10 rows and 50 columns
 b2 = as.data.frame(matrix(rnorm(800),10,80)) # 10 rows and 80 columns
 b3 = new("MultiBlock", Samples = 1:10, Data = list(s1 = b1, s2 = b1),
 Variables = list(s1 = 1:ncol(b1), s2 = letters[1:length(b1)]))
 # Build MultiBlock by adding one data block at a time:
 mb <- BuildMultiBlock(b1, b2, b3)
}
