% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComDim-classes.R
\name{checkMultiBlock}
\alias{checkMultiBlock}
\title{checkMultiBlock}
\usage{
checkMultiBlock(object)
}
\arguments{
\item{object}{A MultiBlock object.}
}
\value{
TRUE if valid, or a character vector of error messages.
}
\description{
Validation function used during MultiBlock building.
}
\keyword{internal}
