% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComDim_PCA.R, R/ComDim_PLS.R
\name{ColumnsPartition}
\alias{ColumnsPartition}
\title{ColumnsPartition}
\usage{
ColumnsPartition(Xn = Xn, Partitions = Partitions)

ColumnsPartition(Xn = Xn, Partitions = Partitions)
}
\arguments{
\item{Xn}{A block.}

\item{Partitions}{The number of partitions.}
}
\value{
A list of column-partitioned sub-matrices.

A list of column-partitioned sub-matrices.
}
\description{
Calculate vertical partitions in a block. Internal function of ComDim_PCA().

Calculate vertical partitions in a block. Internal function of ComDim_PLS().
}
\keyword{internal}
