% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComDim-classes.R
\docType{class}
\name{ComDim-classes}
\alias{ComDim-classes}
\alias{MultiBlock}
\alias{ComDim}
\title{MultiBlock object}
\description{
Object of the type \code{MultiBlock}, to use as input for ComDim
analyses.

The output of a ComDim analysis.
}
\section{Slots}{

\describe{
\item{\code{Samples}}{vector with the sample names. If this data is not available,
the slot will be filled with integers.}

\item{\code{Data}}{A list with the data-blocks.}

\item{\code{Variables}}{A character vector with the variable names. If this data is
not available, the slot will be filled with integers.}

\item{\code{Batch}}{A list with the vectors with the batch information for each
data-block. Facultative.}

\item{\code{Metadata}}{A list with the samples metadata.}

\item{\code{Method}}{The algorithm used in the core of the ComDim analysis (ex. PCA,
PLS,...)}

\item{\code{ndim}}{The number of components.}

\item{\code{Q.scores}}{The Global scores.}

\item{\code{T.scores}}{The Local scores.}

\item{\code{P.loadings}}{The Loadings}

\item{\code{Saliences}}{The Saliences}

\item{\code{R2X}}{The explained variance of the MultiBlock.}

\item{\code{R2Y}}{For regression or discriminant models, the explained variance of
the Y-block.}

\item{\code{Q2}}{For regression or discriminant models, the predicted variance of
the Y-block.}

\item{\code{DQ2}}{For discriminant models, the predicted discriminant variance of the
Y-block.}

\item{\code{Singular}}{The singular values.}

\item{\code{Mean}}{The mean values for each variable in the MultiBlock.}

\item{\code{Norm}}{The norm values for each variable in the MultiBlock.}

\item{\code{PLS.model}}{For ComDim analyses using PLS as core algorithm, it contains
the W, B, B0 and Y).}

\item{\code{cv}}{For ComDim_KOPLS, it contains the index of the samples used during
the cross-validation.}

\item{\code{Prediction}}{A list with the predicted Y, the decision rule used for
sample classification, the sensitivity, the specificity, and a confusion
matrix.}

\item{\code{Metadata}}{vector with the sample names. If this data is not available,
the slot will be filled with integers.}

\item{\code{variable.block}}{A vector with the same length than the P.loadings
indicating the block each variable belongs to.}

\item{\code{runtime}}{The used running time.}
}}

