% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComDim_PCA.R, R/ComDim_PLS.R
\name{Compress_Data_2020}
\alias{Compress_Data_2020}
\title{Compress_Data_2020}
\usage{
Compress_Data_2020(s_n = s_n, CompMethod = CompMethod, Partitions = Partitions)

Compress_Data_2020(s_n = s_n, CompMethod = CompMethod, Partitions = Partitions)
}
\arguments{
\item{s_n}{The multi-block object.}

\item{CompMethod}{It can be 'Normal' (default), 'Kernel', 'PCT', 'Tall' or 'Wide'.}

\item{Partitions}{The number of partitions.}
}
\value{
The compressed multi-block.

The compressed multi-block.
}
\description{
Compress large multi-block objects. Internal function of ComDim_PCA().

Compress large multi-block objects. Internal function of ComDim_PLS().
}
\references{
Original MATLAB implementation:
  \url{https://github.com/DNRutledge/ComDim/blob/main/Compress_Data_2020.m}

Original MATLAB implementation:
  \url{https://github.com/DNRutledge/ComDim/blob/main/Compress_Data_2020.m}
}
\keyword{internal}
