% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-extra.R
\docType{data}
\name{extra}
\alias{extra}
\title{Extracellular metabolites in growth medium}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 12 rows and 298 columns.
}
\source{
\doi{10.15252/msb.202010141}
}
\usage{
data(mouse_ds)
}
\description{
Extracellular metabolites in growth medium
}
\examples{
data(mouse_ds)
# MB <- MultiBlock(Data = list(RNAseq3 = RNAseq3,
#   lipids = lipids, intra = intra, extra = extra))
}
\author{
Radic Shechter et al. (2021) Molecular Systems Biology 17:e10141
(\doi{10.15252/msb.202010141})
}
\keyword{dataset}
