% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBlockNames.R
\name{getBlockNames}
\alias{getBlockNames}
\title{ComDim - Finding common dimensions in multi-block datasets}
\usage{
getBlockNames(MB, DataOrBatchOrMetadata)
}
\arguments{
\item{MB}{A MultiBlock object.}

\item{DataOrBatchOrMetadata}{A string with either "Data", "Batch" or
"Metadata". Facultative.}
}
\value{
A vector with the block names.
}
\description{
Return the block names.
}
