% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleNames.R
\name{getSampleNames}
\alias{getSampleNames}
\title{ComDim - Finding common dimensions in multi-block datasets}
\usage{
getSampleNames(MB)
}
\arguments{
\item{MB}{A MultiBlock object.}
}
\value{
A vector with the Sample names.
}
\description{
Get the sample names.
}
