% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariableNames.R
\name{getVariableNames}
\alias{getVariableNames}
\title{ComDim - Finding common dimensions in multi-block datasets}
\usage{
getVariableNames(MB, block)
}
\arguments{
\item{MB}{A MultiBlock object.}

\item{block}{A vector with the list of blocks.}
}
\value{
The vector with the variable names.
}
\description{
Get the variable names.
}
