% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-KEGG_table_metabolites.R
\docType{data}
\name{KEGG_table_metabolites}
\alias{KEGG_table_metabolites}
\title{Metadata for the metabolites in growth medium}
\format{
An object of class \code{data.frame} with 687 rows and 2 columns.
}
\source{
\href{https://doi.org/10.15252/msb.202010141}{DOI}
}
\usage{
data(mouse_ds)
}
\description{
Metadata for the metabolites in growth medium
}
\examples{
data(mouse_ds)
#' intraMB <- ExpandMB(data = intra, metadata = KEGG_table_metabolites,
#'   minblock = 10, loquace = FALSE)
}
\author{
Radic Shechter et al. (2021) Molecular Systems Biology 17:e10141
(\href{https://doi.org/10.15252/msb.202010141}{DOI})
}
\keyword{dataset}
