% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-lipids.R
\docType{data}
\name{lipids}
\alias{lipids}
\title{Lipid profiles of cell extracts}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 437 rows and 12 columns.
}
\source{
\href{https://doi.org/10.15252/msb.202010141}{DOI}
}
\usage{
data(mouse_ds)
}
\description{
Lipid profiles of cell extracts
}
\examples{
data(mouse_ds)
#' MB <- BuildMultiBlock(t(RNAseq3[,1:12]), t(lipids), t(intra), t(extra))
}
\author{
Radic Shechter et al. (2021) Molecular Systems Biology 17:e10141
(\href{https://doi.org/10.15252/msb.202010141}{DOI})
}
\keyword{dataset}
