% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-metadata_lipids.R
\docType{data}
\name{metadata_lipids}
\alias{metadata_lipids}
\title{Metadata for the lipid profiles of cell extracts}
\format{
An object of class \code{data.frame} with 437 rows and 2 columns.
}
\source{
\doi{10.15252/msb.202010141}
}
\usage{
data(mouse_ds)
}
\description{
Metadata for the lipid profiles of cell extracts
}
\examples{
data(mouse_ds)
#' lipidsMB <- ExpandMB(data = lipids, metadata = metadata_lipids,
#'   minblock = 0, loquace = FALSE)
}
\author{
Radic Shechter et al. (2021) Molecular Systems Biology 17:e10141
(\doi{10.15252/msb.202010141})
}
\keyword{dataset}
