% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiBlock2Matrix.R
\name{MultiBlock2Matrix}
\alias{MultiBlock2Matrix}
\title{MultiBlock2Matrix}
\usage{
MultiBlock2Matrix(MB = MB, blocks = NULL, vars = NULL)
}
\arguments{
\item{MB}{The MultiBlocks structure.}

\item{blocks}{The blocks to combine into a matrix. It can be a vector of integers or a vector with the block names.}

\item{vars}{The variables to kept. It is a list object with the same length as 'blocks'. Each element in the list contains the position of the variables in the nth block.}
}
\value{
The MultiBlock
}
\description{
Combines the blocks of a MultiBlock into a matrix. Useful for calculating summary information of the whole MB (ex. maximum value).
}
\examples{
b1 <- matrix(rnorm(500), 10, 50)
b2 <- matrix(rnorm(800), 10, 80)
mb <- MultiBlock(Data = list(b1 = b1, b2 = b2))
# Convert data to a matrix:
mb <- MultiBlock2Matrix(mb)
max(mb)
}
