% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComDim-classes.R
\name{ncol,MultiBlock-method}
\alias{ncol,MultiBlock-method}
\title{ncol for MultiBlock}
\usage{
\S4method{ncol}{MultiBlock}(x)
}
\arguments{
\item{x}{A MultiBlock object.}
}
\value{
A named integer vector with the number of columns per block.
}
\description{
Return the number of columns (variables) in each block of a MultiBlock.
}
\examples{
b1 <- matrix(rnorm(500), 10, 50)
b2 <- matrix(rnorm(800), 10, 80)
mb <- MultiBlock(Data = list(b1 = b1, b2 = b2))
ncol(mb) # c(b1 = 50, b2 = 80)
}
