% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncolMultiBlock.R
\name{ncolMultiBlock}
\alias{ncolMultiBlock}
\title{ComDim - Finding common dimensions in multi-block datasets}
\usage{
ncolMultiBlock(MB, block)
}
\arguments{
\item{MB}{A MultiBlock.}

\item{block}{The blocks to calculate the number of columns.}
}
\value{
A vector with the number of columns.
}
\description{
Return the number of columns in each block of the MultiBlock.
}
