% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrowMultiBlock.R
\name{nrowMultiBlock}
\alias{nrowMultiBlock}
\title{ComDim - Finding common dimensions in multi-block datasets}
\usage{
nrowMultiBlock(MB, block)
}
\arguments{
\item{MB}{A MultiBlock.}

\item{block}{The blocks to calculate the number of rows}
}
\value{
A vector with the number of rows
}
\description{
Return the number of rows in each block of the MultiBlock.
}
