% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComDim_PCA.R, R/ComDim_PLS.R
\name{PCA_Tall_PCT_DNR}
\alias{PCA_Tall_PCT_DNR}
\title{PCA_Tall_PCT_DNR}
\usage{
PCA_Tall_PCT_DNR(Xn = Xn, PCs = PCs, Partitions = Partitions)

PCA_Tall_PCT_DNR(Xn = Xn, PCs = PCs, Partitions = Partitions)
}
\arguments{
\item{Xn}{The block to compress.}

\item{PCs}{Number of PCs to keep.}

\item{Partitions}{The number of partitions.}
}
\value{
The compressed block.

The compressed block.
}
\description{
Compress a block using the PCT method. Internal function of ComDim_PCA().

Compress a block using the PCT method. Internal function of ComDim_PLS().
}
\keyword{internal}
