% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict_MB.R
\name{Predict_MB}
\alias{Predict_MB}
\title{ComDim - Finding common dimensions in multi-block datasets}
\usage{
Predict_MB(MB = MB, y, model = model, normalise = FALSE, loquace = TRUE)
}
\arguments{
\item{MB}{A MultiBlock object for which prediction is desired.}

\item{y}{An y block to calculate DQ2 in ComDim-PLS and ComDim-OPLS models (optional).}

\item{model}{The ComDim model used for prediction.}

\item{normalise}{To apply normalisation. FALSE == no (default), TRUE == yes.}

\item{loquace}{If TRUE, it returns a message with the ComDim elements to be predicted (i.e. Q.scores, T.scores...)}
}
\value{
A ComDim object.
}
\description{
Predict_MB applies an existent ComDim model to a MultiBlock object.
}
