% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.ComDim-package.R
\docType{package}
\name{R.ComDim-package}
\alias{R.ComDim}
\alias{R.ComDim-package}
\title{R.ComDim: ComDim Analysis in R}
\description{
Provides functions to run ComDim analysis. ComDim is an unsupervised
multi-block method that simultaneously considers multiple data tables to
find latent components common to all tables as well as those specific to
each table.
}
\author{
\strong{Maintainer}: Francesc Puig-Castellví \email{puig.francesc@gmail.com}

}
\keyword{internal}
