% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-rnaseq.R
\docType{data}
\name{rnaseq}
\alias{rnaseq}
\title{RNAseq data on cell differentiation}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 36 rows and 12762 columns.
}
\source{
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE75417}{NCBI}
}
\usage{
data(dataset3)
}
\description{
RNAseq data on cell differentiation
}
\examples{
data(dataset3)
#' mb_d3 <- BuildMultiBlock(rnaseq)
}
\author{
Cabrero et al. (2019) Scientific data 6:256
(\href{https://doi.org/10.1038/s41597-019-0202-7}{DOI})
}
\keyword{dataset}
