% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-RNAseq3.R
\docType{data}
\name{RNAseq3}
\alias{RNAseq3}
\title{RNAseq data of cell extracts}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 12 rows and 9254 columns.
}
\source{
\href{https://seqout.org/p/E-MTAB-8834}{SEQOUT}
}
\usage{
data(mouse_ds)
}
\description{
RNAseq data of cell extracts
}
\examples{
data(mouse_ds)
# MB <- MultiBlock(Data = list(RNAseq3 = RNAseq3,
#   lipids = lipids, intra = intra, extra = extra))
}
\author{
Radic Shechter et al. (2021) Molecular Systems Biology 17:e10141
(\doi{10.15252/msb.202010141})
}
\keyword{dataset}
