% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComDim_PCA.R, R/ComDim_PLS.R
\name{RowsPartition}
\alias{RowsPartition}
\title{RowsPartition}
\usage{
RowsPartition(Xn = Xn, Partitions = Partitions)

RowsPartition(Xn = Xn, Partitions = Partitions)
}
\arguments{
\item{Xn}{A block.}

\item{Partitions}{The number of partitions.}
}
\value{
A list of row-partitioned sub-matrices.

A list of row-partitioned sub-matrices.
}
\description{
Calculate horizontal partitions in a block. Internal function of ComDim_PCA().

Calculate horizontal partitions in a block. Internal function of ComDim_PLS().
}
\keyword{internal}
