% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComDim-classes.R
\name{sampleNames}
\alias{sampleNames}
\alias{sampleNames,MultiBlock-method}
\title{sampleNames}
\usage{
sampleNames(x)

\S4method{sampleNames}{MultiBlock}(x)
}
\arguments{
\item{x}{A MultiBlock object.}
}
\value{
A vector with the sample names.
}
\description{
Return the sample names of a MultiBlock object.
}
\examples{
b1 <- matrix(rnorm(500), 10, 50)
rownames(b1) <- paste0("s", 1:10)
mb <- MultiBlock(Data = list(b1 = b1))
sampleNames(mb) # "s1" ... "s10"
}
