% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComDim-classes.R
\name{sampleNames<-}
\alias{sampleNames<-}
\alias{sampleNames<-,MultiBlock-method}
\title{sampleNames<-}
\usage{
sampleNames(x) <- value

\S4method{sampleNames}{MultiBlock}(x) <- value
}
\arguments{
\item{x}{A MultiBlock object.}

\item{value}{A vector of new sample names. Must have the same length as
the current number of samples.}
}
\value{
The updated MultiBlock object.
}
\description{
Set the sample names of a MultiBlock object.
}
\examples{
b1 <- matrix(rnorm(500), 10, 50)
mb <- MultiBlock(Data = list(b1 = b1))
sampleNames(mb) <- paste0("patient_", 1:10)
sampleNames(mb)
}
