% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBlockNames.R
\name{setBlockNames}
\alias{setBlockNames}
\title{ComDim - Finding common dimensions in multi-block datasets}
\usage{
setBlockNames(MB, block.names)
}
\arguments{
\item{MB}{A MultiBlock object.}

\item{block.names}{A vector with the block names.}
}
\value{
The MultiBlock object.
}
\description{
Set the block names.
}
