% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSampleNames.R
\name{setSampleNames}
\alias{setSampleNames}
\title{ComDim - Finding common dimensions in multi-block datasets}
\usage{
setSampleNames(MB, sample.names)
}
\arguments{
\item{MB}{A MultiBlock object.}

\item{sample.names}{A vector with the new sample names.}
}
\value{
The MultiBlock object.
}
\description{
Get the sample names.
}
