% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setVariableNames.R
\name{setVariableNames}
\alias{setVariableNames}
\title{ComDim - Finding common dimensions in multi-block datasets}
\usage{
setVariableNames(MB, newVars, block)
}
\arguments{
\item{MB}{A MultiBlock object.}

\item{newVars}{A vector with the new variable names.}

\item{block}{A vector with the list of blocks.}
}
\value{
The MultiBlock object
}
\description{
Get the variable names.
}
