% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitRW.R
\name{SplitRW}
\alias{SplitRW}
\title{SplitRW}
\usage{
SplitRW(
  MB = MB,
  checkSampleCorrespondence = FALSE,
  batchNormalisation = TRUE,
  showSampleCorrespondence = TRUE
)
}
\arguments{
\item{MB}{The multi-block structure built with \code{MultiBlock()}.}

\item{checkSampleCorrespondence}{If FALSE, the same number of samples and the sample order are assumed for all the batches. If TRUE, only the samples found in all replicate blocks will be included in the final structure.}

\item{batchNormalisation}{If TRUE, each RW block is divided by its norm and the square root of the number of replicates.}

\item{showSampleCorrespondence}{If TRUE, print in the console a data frame with the sample correspondence across blocks.}
}
\value{
The list containing the multi-block composed of the replicate-blocks.
}
\description{
To split a multi-block into a list containing smaller blocks, comprising each data from one batch.
}
\examples{
b1 <- matrix(rnorm(1500), 30, 50)
b2 <- matrix(rnorm(2400), 30, 80)
batch_b <- c(rep(1, 10), rep(2, 10), rep(3, 10))
# Generate the multi-block (mb) with 3 batches of 10 samples each
mb <- MultiBlock(
  Data = list(b1 = b1, b2 = b2),
  Batch = list(b1 = batch_b, b2 = batch_b),
  ignore.names = TRUE
)
rw <- SplitRW(mb)
}
