% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitVariablesMB.R
\name{SplitVariablesMB}
\alias{SplitVariablesMB}
\title{Internal function used for ExpandMultiBlock
Splits a data.frame or a matrix into smaller sets and returns a MultiBlock object.}
\usage{
SplitVariablesMB(data = NULL, metadata = NULL, columns = NULL, minblock = 0)
}
\arguments{
\item{data}{A data.frame or a matrix. Samples are in rows and variables in columns.}

\item{metadata}{The metadata associated with the variables. It has the same number of rows than columns are in data.}

\item{columns}{The columns from metadata that will guide the split. Optional}

\item{minblock}{The minimal number of variables that need to be contained in a block in order to be kept.}
}
\value{
The MultiBlock.
}
\description{
Internal function used for ExpandMultiBlock
Splits a data.frame or a matrix into smaller sets and returns a MultiBlock object.
}
