% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComDim-classes.R
\name{checkMultiBlock}
\alias{checkMultiBlock}
\title{ComDim - Finding common dimensions in multi-block datasets}
\usage{
checkMultiBlock(object)
}
\arguments{
\item{object}{A MultiBlock object.}
}
\value{
The building of the MultiBlock is interrupted if an error is found.
}
\description{
Validation function used during MultiBlock building.
}
