% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpandMultiBlock.R
\name{ExpandMultiBlock}
\alias{ExpandMultiBlock}
\title{ExpandMultiBlock
Splits data into several blocks, allowing variables to appear on more than one block at the same time.}
\usage{
ExpandMultiBlock(data = NULL, metadata = NULL, minblock = 0, loquace = TRUE)
}
\arguments{
\item{data}{A data.frame or a matrix. Samples are in rows and variables in columns.}

\item{metadata}{A 2-column data.frame detailing how columns in data must be propagated. Entries in the first column must match the row names in data (variable  names). The second column details the new blocks.}

\item{minblock}{The minimal number of variables that need to be contained in a block in order to be kept.}

\item{loquace}{To display warning}
}
\value{
The MultiBlock.
}
\description{
ExpandMultiBlock
Splits data into several blocks, allowing variables to appear on more than one block at the same time.
}
\examples{
data(mouse_ds)
lipidsMB <- ExpandMB(data = lipids, metadata = metadata_lipids,
  minblock = 0, loquace = FALSE)
}
