% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-extra.R
\docType{data}
\name{extra}
\alias{extra}
\title{Extracellular metabolites in growth medium}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 298 rows and 12 columns.
}
\source{
\href{https://doi.org/10.15252/msb.202010141}{DOI}
}
\usage{
data(mouse_ds)
}
\description{
Extracellular metabolites in growth medium
}
\examples{
data(mouse_ds)
#' MB <- BuildMultiBlock(t(RNAseq3[,1:12]), t(lipids), t(intra), t(extra))
}
\author{
Radic Shechter et al. (2021) Molecular Systems Biology 17:e10141
(\href{https://doi.org/10.15252/msb.202010141}{DOI})
}
\keyword{dataset}
