% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterSamplesMultiBlock.R
\name{FilterSamplesMultiBlock}
\alias{FilterSamplesMultiBlock}
\title{ComDim - Finding common dimensions in multi-block datasets}
\usage{
FilterSamplesMultiBlock(MB = MB, samples = getSampleNames(MB))
}
\arguments{
\item{MB}{A MultiBlock object for which prediction is desired.}

\item{samples}{A vector containing the names of the samples to keep.}
}
\value{
A ComDim object.
}
\description{
Predict_MB applies an existent ComDim model to a MultiBlock object.
}
