% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeComDimLoadingsTable.R
\name{MakeComDimLoadingsTable}
\alias{MakeComDimLoadingsTable}
\title{MakeComDimLoadingsTable}
\usage{
MakeComDimLoadingsTable(model, blocks = NULL, dim = NULL, dim.ort = NULL)
}
\arguments{
\item{model}{The output from a ComDim analysis.}

\item{blocks}{The blocks from which the data will be extracted. It can be a vector of integers or a vector with the block names. Optional.}

\item{dim}{The dimensions from which the data will be extracted. It can be a vector of integers. Optional.}

\item{dim.ort}{The orthogonal dimensions from which the data will be extracted. It can be a vector of integers. If NULL, all orthogonal components are extracted. If 0, none of the orthogonal components are extracted. Optional.}
}
\value{
A table with the selected information
}
\description{
Creates a long data frame with the loadings from a ComDim model, suitable for ggplot.
}
\examples{
b1 <- matrix(rnorm(500), 10, 50) # 10 rows and 50 columns
b2 <- matrix(rnorm(800), 10, 80) # 10 rows and 80 columns
mb <- MultiBlock(Data = list(b1 = b1, b2 = b2))
model <- ComDim_PCA(mb, ndim = 2)
tbl <- MakeComDimLoadingsTable(model)
}
