% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment2MultiBlock.R
\name{MultiAssayExperiment2MultiBlock}
\alias{MultiAssayExperiment2MultiBlock}
\title{MultiAssayExperiment2MultiBlock}
\usage{
MultiAssayExperiment2MultiBlock(se, colData_samplenames = NULL, Batch = NULL)
}
\arguments{
\item{se}{The SummarizedExperiment .}

\item{colData_samplenames}{Character string with the column metadata. If colData_samplenames is NULL, the metadata will not be stored.}

\item{Batch}{The name of the column that will be considered as Batch.}
}
\value{
The MultiBlock.
}
\description{
Converts a MultiAssayExperiment2MultiBlock into a MultiBlock.
}
