% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiBlock2MultiAssayExperiment.R
\name{MultiBlock2MultiAssayExperiment}
\alias{MultiBlock2MultiAssayExperiment}
\title{MultiBlock2MultiAssayExperiment}
\usage{
MultiBlock2MultiAssayExperiment(MB, MSEmetadata = NULL)
}
\arguments{
\item{MB}{The MultiBlock object.}

\item{MSEmetadata}{List of unstructured metadata describing the overall content of the object.}
}
\value{
The MultiAssayExperiment object.
}
\description{
Converts MultiBlock into a MultiAssayExperiment.
}
