% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComDim_PCA_MB.R, R/ComDim_PLS_MB.R
\name{PCA_Tall_PCT_DNR}
\alias{PCA_Tall_PCT_DNR}
\title{Compress the multi-block using the PCT method.}
\usage{
PCA_Tall_PCT_DNR(Xn = Xn, PCs = PCs, Partitions = Partitions)

PCA_Tall_PCT_DNR(Xn = Xn, PCs = PCs, Partitions = Partitions)
}
\arguments{
\item{Xn}{The block to compress.}

\item{PCs}{Number of PCs to keep.}

\item{Partitions}{The number of partitions.}
}
\value{
The compressed multi-block.

The compressed multi-block.
}
\description{
Internal function of ComDim_PCA().

Internal function of ComDim_PCA().
}
\examples{
b1 <- matrix(rnorm(5000),10,500)
b1 <- PCA_Tall_PCT_DNR(b1,10,2)
b1 <- matrix(rnorm(5000),10,500)
b1 <- PCA_Tall_PCT_DNR(b1,10,2)
}
