% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-RNAseq3.R
\docType{data}
\name{RNAseq3}
\alias{RNAseq3}
\title{RNAseq data of cell extracts}
\format{
An object of class \code{data.frame} with 9254 rows and 13 columns.
}
\source{
\href{https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-8834/}{ARRAYEXPRESS}
}
\usage{
data(RNAseq3)
}
\description{
RNAseq data of cell extracts
}
\examples{
data(mouse_ds)
#' MB <- BuildMultiBlock(t(RNAseq3[,1:12]), t(lipids), t(intra), t(extra))
}
\author{
Radic Shechter et al. (2021) Molecular Systems Biology 17:e10141
(\href{https://doi.org/10.15252/msb.202010141}{DOI})
}
\keyword{dataset}
