% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectFeaturesRW.R
\name{SelectFeaturesRW}
\alias{SelectFeaturesRW}
\title{SelectFeaturesRW}
\usage{
SelectFeaturesRW(
  RW = RW,
  results = results,
  ndim = NULL,
  blocks = NULL,
  threshold_cor = 1,
  threshold_cov = 1,
  mean.RW = T,
  plots = "NO"
)
}
\arguments{
\item{RW}{The object used as input in the ComDim analysis.}

\item{results}{The output object obtained in the ComDim analysis.}

\item{ndim}{The number of the component for which the most important variables are to be calculated.}

\item{blocks}{A vector with the indices or the names for the replicate blocks of the same data type.}

\item{threshold_cor}{The "times" parameter used to calculate the threshold in the following formula: cor(variable) > times * sd(cor(variables)). Minimal value that can be assigned to threshold_cor is 1.}

\item{threshold_cov}{The "times" parameter used to calculate the threshold in the following formula: cov(variable) > times * sd(cov(variables)). Minimal value that can be assigned to threshold_cor is 1.}

\item{mean.RW}{Logical value to indicate whether the RW data must be mean-centered (TRUE) or not (FALSE).}

\item{plots}{Parameter to indicate whether a plot must be produced. Possible values are "NO" for no plots, "separated" for plotting each plot individually, and "together" to plot all the plots in the same grid.}
}
\value{
An object with 2 lists. The first list contains the important variables presenting a positive relationship with the scores, while the second list contains the most important variables presenting a negative relationship.
}
\description{
Finds the important variables presenting a coordinated response in all the replicate-blocks
}
\examples{
 b1 = matrix(rnorm(500),10,50)
 batch_b1 = rep(1,10)
 b2 = matrix(rnorm(800),30,80)
 batch_b2 = c(rep(1,10),rep(2,10),rep(3,10))
 mb <- BuildMultiBlock(b1, batches = batch_b1)
 mb <- BuildMultiBlock(b2, growingMB = mb, batches = batch_b2, equalSampleNumber = FALSE)
 rw <- SplitRW(mb)
 results<-ComDim_PCA(rw, 2) # In this analysis, we used 2 components.
 features <- SelectFeaturesRW(RW = rw, results = results, ndim = 1, blocks = c(2,3,4))
}
