% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedExperiment2MultiBlock.R
\name{SummarizedExperiment2MultiBlock}
\alias{SummarizedExperiment2MultiBlock}
\title{SummarizedExperiment2MultiBlock}
\usage{
SummarizedExperiment2MultiBlock(se, colData_samplenames = NULL, Batch = NULL)
}
\arguments{
\item{se}{The SummarizedExperiment .}

\item{colData_samplenames}{Character string with the column metadata.}

\item{Batch}{The name of the column that will be considered as Batch.}
}
\value{
The MultiBlock
}
\description{
Converts a SummarizedExperiment into a MultiBlock.
}
